==> RecSys2020_Challenge/01_Preprocess/CPU/6-build_sample_submission_files.ipynb <==
Copyright (c) 2020, NVIDIA CORPORATION.
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

import matplotlib.pyplot as plt
%matplotlib inline  

import pandas as pd
import numpy as np
import gc

DTYPES = {
    '8': 'int32',
    '10': 'int32',
    '11': 'int32',
    '13': 'int32',
    '15': 'int32',
    '16': 'int32',
    '18': 'int32',
    '20': 'float32',
    '21': 'float32',
    '22': 'float32',
    '23': 'float32',
}

dv = pd.read_csv('../input/val.tsv', sep='\x01', header=None, dtype=DTYPES )
gc.collect()
dv.shape

features = [
    'text_tokens',    ###############
    'hashtags',       #Tweet Features
    'tweet_id',       #
    'media',          #
    'links',          #
    'domains',        #
    'tweet_type',     #
    'language',       #
    'timestamp',      ###############
    'a_user_id',              ###########################
    'a_follower_count',       #Engaged With User Features
    'a_following_count',      #
    'a_is_verified',          #
    'a_account_creation',     ###########################
    'b_user_id',              #######################
    'b_follower_count',       #Engaging User Features
    'b_following_count',      #
    'b_is_verified',          #
    'b_account_creation',     #######################
    'b_follows_a',    #################### Engagement Features
    #'reply',          #Target Reply
    #'retweet',        #Target Retweet    
    #'retweet_comment',#Target Retweet with comment
    #'like',           #Target Like
                      ####################
]
dv.columns = features
gc.collect()
dv.head()

dv = dv[['tweet_id','b_user_id']].copy()
dv['prediction'] = 0
dv.to_csv( 'sample_submission_public.csv', index=False )
dv.head()

dv = pd.read_csv('../input/competition_test.tsv', sep='\x01', header=None, dtype=DTYPES )
dv.columns = features
gc.collect()
print( dv.shape )
dv.head()

dv = dv[['tweet_id','b_user_id']].copy()
dv['prediction'] = 0
dv.to_csv( 'sample_submission_private.csv', index=False )
dv.head()
